requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
google-api-core==1.26.0
google-auth==1.27.0
google-cloud==0.34.0
google-cloud-vision==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
libcst==0.3.17
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
gunicorn==20.0.4
whitenoise==5.2.0
